JPA Test

+++ src/main/java/com/demoon/uplearning/Service/ApplicantService.java
@@ -4,5 +4,5 @@ package com.demoon.uplearning.Service;
 import com.demoon.uplearning.entity.User;
 
 public interface ApplicantService {
-    public User getUserByID(Long id);
+    public User getUserByID(Integer id);
 }
